+++ src/include/poolcommon/multiCall.h
@@ -7,15 +7,15 @@
 template<typename T>
 struct MultiCall {
   std::unique_ptr<T[]> Data;
-  std::atomic<uint32_t> FinishedCallsNum = 0;
-  uint32_t TotalCallsNum;
+  std::atomic<size_t> FinishedCallsNum = 0;
+  size_t TotalCallsNum;
   std::function<void(const T*, size_t)> MainCallback;
 
-  MultiCall(uint32_t totalCallsNum, std::function<void(const T*, size_t)> mainCallback) : TotalCallsNum(totalCallsNum), MainCallback(mainCallback) {
+  MultiCall(size_t totalCallsNum, std::function<void(const T*, size_t)> mainCallback) : TotalCallsNum(totalCallsNum), MainCallback(mainCallback) {
     Data.reset(new T[totalCallsNum]);
   }
 
-  std::function<void(const T&)> generateCallback(uint32_t callNum) {
+  std::function<void(const T&)> generateCallback(size_t callNum) {
     return [this, callNum](const T &data) {
       Data[callNum] = data;
       if (++FinishedCallsNum == TotalCallsNum) {
